+++ src/migrations/M191106150324Slugs.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace davidhirtz\yii2\cms\migrations;
 
 use davidhirtz\yii2\cms\migrations\traits\I18nTablesTrait;
+use davidhirtz\yii2\cms\migrations\traits\SlugIndexTrait;
 use davidhirtz\yii2\cms\models\Category;
-use davidhirtz\yii2\cms\models\Entry;
 use davidhirtz\yii2\skeleton\db\traits\MigrationTrait;
 use Exception;
 use yii\db\Migration;
@@ -19,6 +19,7 @@ class M191106150324Slugs extends Migration
 {
     use MigrationTrait;
     use I18nTablesTrait;
+    use SlugIndexTrait;
 
     public function safeUp(): void
     {
@@ -27,21 +28,15 @@ class M191106150324Slugs extends Migration
 
             foreach ($category->getI18nAttributeNames('slug') as $attributeName) {
                 try {
-                    $this->dropIndex($attributeName, Category::tableName());
-                    $this->createIndex($attributeName, Category::tableName(), $category->slugTargetAttribute ?: $attributeName, true);
+                    $this->dropIndex($attributeName, $category::tableName());
                 } catch (Exception) {
                 }
+
+                $this->createIndex($attributeName, $category::tableName(), $category->slugTargetAttribute ?: $attributeName, true);
             }
 
-            $entry = Entry::instance();
-
-            foreach ($entry->getI18nAttributeNames('slug') as $attributeName) {
-                try {
-                    $this->dropIndex($attributeName, Entry::tableName());
-                    $this->createIndex($attributeName, Entry::tableName(), $entry->slugTargetAttribute ?: $attributeName, true);
-                } catch (Exception) {
-                }
-            }
+            $this->dropSlugIndex();
+            $this->createSlugIndex();
         });
     }
 }